pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mngs"
dynamic = ["version"]
description = "For lazy python users (monogusa people in Japanese), especially in ML/DSP fields"
authors = [{name = "ywatanabe1989", email = "ywatanabe@alumni.u-tokyo.ac.jp"}]
readme = "README.md"
license = {text = "MIT"}
keywords = ["utils", "utilities", "python", "machine learning"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
requires-python = ">=3.0"
dependencies = [
    "GitPython",
    "Pillow",
    "PyYAML",
    "Pyarrow",
    "accelerate",
    "aiohttp",
    "ansi-escapes",
    "anthropic",
    "attrdict",
    "bitsandbytes",
    "black",
    "bs4",
    "chardet",
    "docutils<0.18",
    "einops",
    "epc",
    "fairscale",
    "fastapi",
    "flake8",
    "flask",
    "geom_median",
    "google-genai",
    "google-search-results",
    "googlesearch-python",
    "groq",
    "h5py",
    "html2text",
    "icecream",
    "ipdb",
    "ipython",
    "isort",
    "jedi",
    "joblib",
    "joypy",
    "julius",
    "llama-stack",
    "lxml",
    "lxml_html_clean",
    "markdown",
    "markdown2",
    "matplotlib",
    "matplotlib-venn",
    "mne",
    "more-itertools",
    "natsort",
    "numpy",
    "obspy",
    "openai",
    "openpyxl",
    "optuna",
    "packaging",
    "pandas",
    "pebble",
    "platformdirs",
    "plotly",
    "plyer",
    "psutil",
    "psycopg2-binary",
    "pybids",
    "pyedflib",
    "pygments",
    "pyls",
    "pymatreader",
    "pyperclip",
    "pyright",
    "pytest",
    "pytest-cov",
    "pytest-env",
    "pytest-xdist",
    "python-docx",
    "python-lsp-server",
    "pytorch_pretrained_vit",
    "pytorch-optimizer",
    "readability",
    "readability-lxml",
    "readchar",
    "reportlab",
    "requests",
    "ripple_detection",
    "ruamel.yaml",
    "ruff-lsp",
    "scikit-image",
    "scikit-learn",
    "scipy",
    "seaborn",
    "selenium",
    "setuptools",
    "six",
    "sktime",
    "sounddevice",
    "sphinx",
    "sqlalchemy",
    "statsmodels",
    "sympy",
    "tabulate",
    "tensorpac",
    "termplotlib",
    "timeout-decorator",
    "tk",
    "tldr",
    "torch",
    "torchaudio",
    "torchsummary",
    "torchvision",
    "transformers",
    "twine",
    "umap-learn",
    "webdriver-manager",
    "wheel",
    "xarray",
    "xlrd",
    "xmltodict",
]

[project.urls]
Homepage = "https://github.com/ywatanabe1989/mngs"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "mngs.__version__"}
